// src/core/interfaces/sku.ts
import { UUID } from "crypto";

interface Item {
    ItemCode?: string;
    ItemName?: string;
    ForeignName?: string;
    ItemsGroupCode?: number;
    InventoryItem?: string;
    SalesItem?: string;
    PurchaseItem?: string;
    ItemType?: string;
    ItemClass?: string;
    BarCode?: string;
    MaterialType?: number;
    NCMCode?: number;
    CESTCode?: number;
    ProductSource?: number;
    PurchaseUnit?: string;
    PurchaseItemsPerUnit?: number;
    PurchasePackagingUnit?: string;
    PurchaseQtyPerPackUnit?: number;
    SalesUnit?: string;
    SalesItemsPerUnit?: number;
    SalesPackagingUnit?: string;
    SalesQtyPerPackUnit?: number;
    Mainsupplier?: string;
    SupplierCatalogNo?: string;
    InventoryUOM?: string;
    SalesUnitLength?: number;
    SalesLengthUnit?: number;
    SalesUnitWidth?: number;
    SalesWidthUnit?: number;
    SalesUnitHeight?: number;
    SalesHeightUnit?: number;
    SalesUnitVolume?: number;
    SalesVolumeUnit?: number;
    SalesUnitWeight?: number;
    SalesWeightUnit?: number;
    PurchaseUnitLength?: number;
    PurchaseLengthUnit?: number;
    PurchaseUnitWidth?: number;
    PurchaseWidthUnit?: number;
    PurchaseUnitHeight?: number;
    PurchaseHeightUnit?: number;
    PurchaseUnitVolume?: number;
    PurchaseVolumeUnit?: number;
    PurchaseUnitWeight?: number;
    PurchaseWeightUnit?: number;
    U_GCV_brand_code?: string;
    U_GCV_collection_code?: string;
    U_GCV_category_code_1?: string;
    U_GCV_category_code_2?: string;
    U_GCV_category_code_3?: string;
    U_GCV_category_code_4?: string;
    U_GCV_category_code_5?: string;
    U_GCV_product_code?: string;
    U_GCV_variant_code?: string;
    U_GCV_color_code?: string;
    U_GCV_grid_code?: string;
    U_GCV_gridItem_prefix?: string;
    ItemPrices?: ItemPrice[];
}

interface ItemPrice {
    PriceList?: number;
    Price?: number;
    Currency?: string;
    BasePriceList?: number;
    Factor?: number;
}

interface SKU {
    ID?: UUID;
    code?: string;
    name?: string;
    barcode?: string;
    color_code?: string;
    grid_code?: string;
    gridItem_prefix?: string;
    itemGroup_code?: number;
    material_code?: number;
    ncm_code?: number;
    source_code?: number;
    cest_code?: number;
    purchaseUnit_code?: string;
    purchaseItemsPerUnit?: number;
    purchasePack_code?: string;
    purchaseQtyPerPackUnit?: number;
    salesUnit_code?: string;
    salesItemsPerUnit?: number;
    salesPack_code?: string;
    salesQtyPerPackUnit?: number;
    vendor_code?: string;
    vendorReference?: string;
    inventoryUOM_code?: string;
    salesUnitLength?: string;
    salesLengthUnit_code?: number;
    salesUnitWidth?: string;
    salesWidthUnit_code?: number;
    salesUnitHeight?: string;
    salesHeightUnit_code?: number;
    salesUnitVolume?: string;
    salesVolumeUnit_code?: number;
    salesUnitWeight?: string;
    salesWeightUnit_code?: number;
    purchaseUnitLength?: string;
    purchaseLengthUnit_code?: number;
    purchaseUnitWidth?: string;
    purchaseWidthUnit_code?: number;
    purchaseUnitHeight?: string;
    purchaseHeightUnit_code?: number;
    purchaseUnitVolume?: string;
    purchaseVolumeUnit_code?: number;
    purchaseUnitWeight?: string;
    purchaseWeightUnit_code?: number;
    brand_code?: string;
    collection_code?: string;
    category_code_1?: string;
    category_code_2?: string;
    category_code_3?: string;
    category_code_4?: string;
    category_code_5?: string;
    product_code?: string;
    variant_code : string
}

export { Item, ItemPrice, SKU };